python-pysqlite:ver.bump renamed
