+++ programming/language/python/python-pysqlite/pspec.xml
@@ -2,37 +2,45 @@
 <!DOCTYPE PISI SYSTEM 'http://www.pisilinux.org/projeler/pisi/pisi-spec.dtd'>
 <PISI>
     <Source>
-        <Name>pysqlite</Name>
-        <Homepage>http://oss.itsystementwicklung.de/trac/pysqlite</Homepage>
+        <Name>python-pysqlite</Name>
+        <Homepage>https://github.com/ghaering/pysqlite</Homepage>
         <Packager>
             <Name>PisiLinux Community</Name>
             <Email>admins@pisilinux.org</Email>
         </Packager>
         <License>PSF-2.2</License>
+        <PartOf>programming.language.python</PartOf>
         <IsA>library</IsA>
         <Summary>Python sqlite bindings</Summary>
         <Description>pysqlite is a DB-API 2.0-compliant database interface for SQLite. pysqlite makes the powerful embedded SQL engine SQLite available to Python programmers. It stays compatible with the Python database API specification 2.0 as much as possible, but also exposes most of SQLite's native API, so that it is for example possible to create user-defined SQL functions and aggregates in Python.</Description>
-        <Archive sha1sum="ea2c7946400305a92b7670ec6758c012490db70f" type="targz">https://pypi.python.org/packages/source/p/pysqlite/pysqlite-2.8.1.tar.gz</Archive>
+        <Archive sha1sum="70e6ed5e771b1b584ffe23f93e3eb1ceef44895a" type="targz">https://pypi.io/packages/source/p/pysqlite/pysqlite-2.8.3.tar.gz</Archive>
         <BuildDependencies>
             <Dependency>sqlite-devel</Dependency>
-            <Dependency>python-sphinx</Dependency>
-	    <Dependency>python-devel</Dependency>
+            <!--<Dependency>python-sphinx</Dependency>-->
+            <Dependency>python-devel</Dependency>
         </BuildDependencies>
     </Source>
 
     <Package>
-        <Name>pysqlite</Name>
-        <RuntimeDependencies>
+        <Name>python-pysqlite</Name>
+        <!--<RuntimeDependencies>
             <Dependency>sqlite</Dependency>
             <Dependency>python</Dependency>
-        </RuntimeDependencies>
+        </RuntimeDependencies>-->
         <Files>
             <Path fileType="library">/usr/lib</Path>
-            <Path fileType="doc">/usr/share/doc</Path>
+            <Path fileType="doc">/usr/share/doc/python-pysqlite</Path>
         </Files>
     </Package>
 
     <History>
+        <Update release="5">
+            <Date>2019-11-21</Date>
+            <Version>2.8.3</Version>
+            <Comment>Ver. bump</Comment>
+            <Name>Blue DeviL</Name>
+            <Email>bluedevil@sctzine.com</Email>
+        </Update>
       <Update release="4">
             <Date>2015-10-10</Date>
             <Version>2.8.1</Version>
