+++ programming/language/python/python-pysqlite/translations.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" ?>
 <PISI>
     <Source>
-        <Name>pysqlite</Name>
+        <Name>python-pysqlite</Name>
         <Summary xml:lang="tr">Python sqlite bağlayıcıları</Summary>
         <Description xml:lang="tr">pysqlite, SQLite için bir DB-API 2.0 uyumlu veri tabanı ara yüzüdür. pysql güçlü gömülü SQL motoru SQLite'ı Python programcılarına uygun hale getirir. Python veri tabanı API 2.0 teknik özelliklerine olabildiğince uyumlu kalır, ama aynı zamanda SQLite'ın doğal API'sinin çoğunu meydana çıkarır, böylece örneğin, Python ile kullanıcı tanımlı SQL fonksiyonları ve kümeler oluşturmak mümkündür.</Description>
         <Description xml:lang="fr">pysqlite est une interface pour bases de données SQLite conforme à DB-API 2.0. pysqlite rend le puissant moteur embarqué SQL SQLite disponible pour les programmeurs Python. Elle demeure aussi compatible que possible avec les spécifications de l'API de bases de données 2.0, mais expose également une large partie de l'API native de SQLite, afin qu'il soit par exemple possible de créer des fonctions et des agrégats (aggregates) SQL utilisateur depuis Python.</Description>
